Refactor CLI structure to unify server and client commands
Consolidated server and client commands under a single "mit" CLI structure for better organization. Moved shared functionality to a unified "cmd" package and adjusted associated tests and imports accordingly. Simplified command initialization to allow hierarchical subcommands for clearer usage.
