+++ pkg/cmd/init_test.go
@@ -1,4 +1,4 @@
-package server
+package cmd
 
 import (
 	"testing"
